update_esame cancella_esame
deletion and update for exam from a teacher finally working as expected

+++ docente/cancella_esame.php
@@ -1,25 +1,29 @@
 <?php
 //obiettivo: cancellare l'esame specificato in $esame
-    $esame_id = $_GET['esame']; //non so perchè devo prelevare da insegnamento
-    $operazione = $_GET['operazione'];
+    $esame_id = $_GET['cancella_esame']; 
     print("esame : ". $esame_id."</br>");
-    print("operazione: ". $operazione);
     
     $db = pg_connect("dbname=unimio host=localhost port=5432");
-            $sql = "DELETE FROM esami 
-                    WHERE id = $1";
-            $preparazione = pg_prepare($db , "cancellazione", $sql);
-            if($preparazione){
-                //pg_execute($db, "cancellazione", array($esame_id));
-                print("cancellazione dell'esame");
-
-                $url_errore ="docente.php?approved=" . urlencode(0) ;
-                header("Location: " . $url_errore);
-                exit;
-            }else{
-                print("qualcosa è andato storto nella preparazione della query");
-                $url_errore ="docente.php?approved=" . urlencode(1) . "&msg="  . urlencode("la cancellazione dell'esame non è andata a buon fine");
-                header("Location: " . $url_errore);
-                exit;
+    if($db){
+        $sql = "DELETE FROM esami 
+                WHERE id = $1";
+        $preparazione = pg_prepare($db , "cancellazione", $sql);
+        if($preparazione){
+            print'query preparata con successo';
+            $eseguito = pg_execute($db, "cancellazione", array($esame_id));
+        
+            if($eseguito){
+                
+                //ritorno un messaggio di successo
+                $_POST['msg'] = "esame". $esame_id ." cancellato con successo";
+                $_POST['approved'] = 0;
             }
+        }else{//ritorno un messaggio di errore
+            $_POST['msg'] = pg_last_error();
+            $_POST['approved'] = 1;
+        }
+    }else{
+        $_POST['msg'] = "connessione al database fallita";
+        $_POST['approved'] = 1;
+    }
 ?>
